refactor: move the specialization of SQL
