+++ api/src/main/java/jakarta/data/repository/jpa/KeysetAwareSlice.java
@@ -15,7 +15,11 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.repository;
+package jakarta.data.repository.jpa;
+
+import jakarta.data.repository.Pageable;
+import jakarta.data.repository.Slice;
+import jakarta.data.repository.Sort;
 
 /**
  * <p>Keyset pagination is a form of pagination that aims to reduce the
